separate out the agent from the environment
